+++ src/index.ts
@@ -1,7 +1,7 @@
 import type { CheckOptions } from './types'
 
 export { resolvePackage, resolveDependencies, resolveDependency } from './io/resolves'
-export { loadPackageJSON as loadPackage, loadPackages, writePackage } from './io/packages'
+export { loadPackage, loadPackages, writePackage } from './io/packages'
 export { dumpDependencies, parseDependencies } from './io/dependencies'
 export { CheckPackages } from './api/check'
 export * from './types'
